requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2==2.8.5
django-filter==2.2.0
django-rest-auth==0.9.5
drf-access-policy==0.5.1
factory-boy==2.12.0